requirements.txt
faiss-cpu
langchain
langchain-community
langchain-google-genai
python-dotenv
google-generativeai
python-dotenv 
